Move SELECT_KERNEL_ID_POSITRON to avoid cyclic dependency errors
